requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
